# contracts/operations/v1/openapi.yaml
openapi: 3.1.0
info:
  title: BoatOps Internal Operations API
  version: 1.0.0-alpha.5
  summary: Private operations commands and operator schedule catalog/calendar reads.
  description: >-
    Private alpha contract. Inventory and trip commands require operations.write.
    Finance writes require operations.finance.write, and finance reads require
    operations.finance.read or operations.finance.write. Schedule reads require
    operations.schedule.read or operations.schedule.write; schedule writes require
    operations.schedule.write. Schedule catalog writes use canonical server-owned
    identities and audit logs. This API is not part of the channel-facing Inventory
    Provider API and must not be exposed to untrusted channel clients.
servers:
  - url: https://boatops.ayany.com
    description: Ayany private alpha target
  - url: http://localhost:8000
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/internal/v1/schedule/slot-offerings:
    get:
      tags: [Schedule]
      operationId: listSlotOfferings
      summary: List the organization slot catalog and canonical compatibility rules
      description: Requires operations.schedule.read or operations.schedule.write.
      responses:
        '200':
          description: Organization-scoped presets, custom templates, instances, status, boat scope, and rules.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SlotCatalogResponse'}
        '403': {$ref: '#/components/responses/AuthorizationFailed'}
    post:
      tags: [Schedule]
      operationId: createSlotOffering
      summary: Create a reusable custom slot definition
      description: Requires operations.schedule.write. Clock times remain unverified until operating rules are frozen.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateSlotOfferingRequest'}
      responses:
        '201':
          description: New immutable-identity reusable definition and audit row created.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SlotOfferingCommandResponse'}
        '403': {$ref: '#/components/responses/AuthorizationFailed'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/internal/v1/schedule/custom-slot-instances:
    post:
      tags: [Schedule]
      operationId: createCustomSlotInstance
      summary: Create a date-specific custom slot instance or template override
      description: >-
        Requires operations.schedule.write. The fictional FULL_DAY_6H 12:00-18:00
        validation scenario must be represented here, not by changing the reusable preset.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateCustomSlotInstanceRequest'}
      responses:
        '201':
          description: Date-specific custom identity and audit row created.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CustomSlotInstanceCommandResponse'}
        '403': {$ref: '#/components/responses/AuthorizationFailed'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
...
  /api/internal/v1/schedule/slot-offerings/{id}:activate:
    post:
      tags: [Schedule]
      operationId: activateSlotOffering
      summary: Transition a DRAFT slot definition to ACTIVE
      description: Requires operations.schedule.write and creates an audit row.
      parameters:
        - $ref: '#/components/parameters/ObjectId'
      responses:
        '200':
          description: Definition activated for future selection.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SlotOfferingCommandResponse'}
        '403': {$ref: '#/components/responses/AuthorizationFailed'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/internal/v1/schedule/slot-offerings/{id}:retire:
    post:
      tags: [Schedule]
      operationId: retireSlotOffering
      summary: Transition a DRAFT or ACTIVE slot definition to RETIRED
      description: >-
        Requires operations.schedule.write and creates an audit row. RETIRED prevents
        future selection but never changes historical HOLD/allocation/booking snapshots.
      parameters:
        - $ref: '#/components/parameters/ObjectId'
      responses:
        '200':
          description: Definition retired without deleting history.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SlotOfferingCommandResponse'}
        '403': {$ref: '#/components/responses/AuthorizationFailed'}
        '409': {$ref: '#/components/responses/Conflict'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque API token with explicit operations, finance, or schedule scope
  parameters:
...
    ObjectId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
    AuthorizationFailed:
      description: Authentication, organization isolation, or required operation scope failed.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    Conflict:
      description: Inventory, state, readiness, external-reference, or idempotency conflict.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    ValidationFailed:
      description: Request validation failed.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
  schemas:
    CreateSlotOfferingRequest:
      type: object
      additionalProperties: false
      required: [code, name, service_start_time, service_end_time, duration_minutes]
      properties:
        code: {type: string, pattern: '^[A-Z0-9][A-Z0-9_-]{1,99}$', maxLength: 100}
        name: {type: string, minLength: 1, maxLength: 255}
        status: {type: string, enum: [DRAFT, ACTIVE], default: DRAFT}
        service_start_time: {type: string, pattern: '^(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d)?$'}
        service_end_time: {type: string, pattern: '^(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d)?$'}
        duration_minutes: {type: integer, minimum: 1, maximum: 1440}
        additional_buffer_before_minutes: {type: integer, minimum: 0, maximum: 1440, default: 0}
        additional_buffer_after_minutes: {type: integer, minimum: 0, maximum: 1440, default: 0}
        valid_from: {type: [string, 'null'], format: date}
        valid_until: {type: [string, 'null'], format: date}
        applies_to_all_boats: {type: boolean, default: true}
        boat_ids:
          type: array
          maxItems: 100
          uniqueItems: true
          items: {type: integer, minimum: 1}
    CreateCustomSlotInstanceRequest:
      type: object
      additionalProperties: false
      required: [code, service_date]
      properties:
        template_slot_offering_id: {type: [integer, 'null'], minimum: 1}
        code: {type: string, pattern: '^[A-Z0-9][A-Z0-9_-]{1,99}$', maxLength: 100}
        name: {type: [string, 'null'], minLength: 1, maxLength: 255}
        status: {type: string, enum: [DRAFT, ACTIVE], default: DRAFT}
        service_date: {type: string, format: date}
        service_start_time: {type: [string, 'null'], pattern: '^(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d)?$'}
        service_end_time: {type: [string, 'null'], pattern: '^(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d)?$'}
        duration_minutes: {type: [integer, 'null'], minimum: 1, maximum: 1440}
        additional_buffer_before_minutes: {type: integer, minimum: 0, maximum: 1440, default: 0}
        additional_buffer_after_minutes: {type: integer, minimum: 0, maximum: 1440, default: 0}
        applies_to_all_boats: {type: boolean, default: true}
        boat_ids:
          type: array
          maxItems: 100
          uniqueItems: true
          items: {type: integer, minimum: 1}
...
    SlotOffering:
      type: object
      additionalProperties: false
      required: [id, template_slot_offering_id, kind, code, name, status, operating_time_status, operating_time_notice, service_date, service_start_time, service_end_time, duration_minutes, additional_buffer_before_minutes, additional_buffer_after_minutes, valid_from, valid_until, applies_to_all_boats, boat_ids, has_historical_usage, created_at, updated_at]
      properties:
        id: {type: integer, minimum: 1}
        template_slot_offering_id: {type: [integer, 'null'], minimum: 1}
        kind: {type: string, enum: [PRESET, CUSTOM_TEMPLATE, CUSTOM_INSTANCE]}
        code: {type: string}
        name: {type: string}
        status: {type: string, enum: [ACTIVE, DRAFT, RETIRED]}
        operating_time_status: {type: string, enum: [UNVERIFIED, DEMO_DEFAULT_UNVERIFIED, FICTIONAL_VALIDATION_SCENARIO]}
        operating_time_notice: {type: string}
        service_date: {type: [string, 'null'], format: date}
        service_start_time: {type: string}
        service_end_time: {type: string}
        duration_minutes: {type: integer, minimum: 1, maximum: 1440}
        additional_buffer_before_minutes: {type: integer, minimum: 0}
        additional_buffer_after_minutes: {type: integer, minimum: 0}
        valid_from: {type: [string, 'null'], format: date}
        valid_until: {type: [string, 'null'], format: date}
        applies_to_all_boats: {type: boolean}
        boat_ids:
          type: array
          uniqueItems: true
          items: {type: integer, minimum: 1}
        has_historical_usage: {type: boolean}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    CompatibilityRule:
      type: object
      additionalProperties: false
      required: [id, pair_key, first_slot_offering_id, first_slot_code, first_slot_name, second_slot_offering_id, second_slot_code, second_slot_name, policy, reason, updated_at]
      properties:
        id: {type: integer, minimum: 1}
        pair_key: {type: string, pattern: '^\d+:\d+$'}
        first_slot_offering_id: {type: integer, minimum: 1}
        first_slot_code: {type: string}
        first_slot_name: {type: string}
        second_slot_offering_id: {type: integer, minimum: 1}
        second_slot_code: {type: string}
        second_slot_name: {type: string}
        policy: {type: string, enum: [ALLOW, DENY]}
        reason: {type: [string, 'null']}
        updated_at: {type: string, format: date-time}
    SlotCatalogResponse:
      type: object
      additionalProperties: false
      required: [request_id, organization_id, inventory_revision, operating_time_notice, slot_offerings, compatibility_rules, code, occurred_at, business_timezone]
      properties:
        request_id: {type: string, format: uuid}
        organization_id: {type: integer, minimum: 1}
        inventory_revision: {type: integer, minimum: 0}
        operating_time_notice: {type: string}
        slot_offerings:
          type: array
          items: {$ref: '#/components/schemas/SlotOffering'}
        compatibility_rules:
          type: array
          items: {$ref: '#/components/schemas/CompatibilityRule'}
        code: {const: SLOT_OFFERINGS_LISTED}
        occurred_at: {type: string, format: date-time}
        business_timezone: {type: string}
    SlotOfferingCommandResponse:
      type: object
      additionalProperties: false
      required: [request_id, organization_id, slot_offering, inventory_revision, code, occurred_at, business_timezone]
      properties:
        request_id: {type: string, format: uuid}
        organization_id: {type: integer, minimum: 1}
        slot_offering: {$ref: '#/components/schemas/SlotOffering'}
        inventory_revision: {type: integer, minimum: 1}
        code: {type: string, enum: [SLOT_OFFERING_CREATED, SLOT_OFFERING_ACTIVATED, SLOT_OFFERING_RETIRED]}
        occurred_at: {type: string, format: date-time}
        business_timezone: {type: string}
    CustomSlotInstanceCommandResponse:
      type: object
      additionalProperties: false
      required: [request_id, organization_id, custom_slot_instance, inventory_revision, code, occurred_at, business_timezone]
      properties:
        request_id: {type: string, format: uuid}
        organization_id: {type: integer, minimum: 1}
        custom_slot_instance: {$ref: '#/components/schemas/SlotOffering'}
        inventory_revision: {type: integer, minimum: 1}
        code: {const: CUSTOM_SLOT_INSTANCE_CREATED}
        occurred_at: {type: string, format: date-time}
        business_timezone: {type: string}
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [request_id, code, message, retryable, manual_action_required]
      properties:
        request_id: {type: string, format: uuid}
        code: {type: string}
        message: {type: string}
        retryable: {type: boolean}
        manual_action_required: {type: boolean}